Categoria insumo ya no requiere imagen, solo un JSON y se implemento la formula para calcular el precio por el tiempo en cocina, se implemento un nuevo controllador para configurar los precios y la cantidad de cocineros, empeze lo de Mercado pago aun sin terminar, Los websockets tambien estan implementados pero no van a implementarse por falta de tiempo

+++ Back-end/src/main/java/com/example/demo/Controllers/DTOS/CarritoDTO.java
@@ -17,7 +17,7 @@ import java.util.List;
 @Setter
 public class CarritoDTO {
     List<ProductosCarritoDTO> productosComprados = new ArrayList<ProductosCarritoDTO>();
-    private  int totalCompra;
+    private  Double totalCompra;
 
 
 }
